Reactのダウングレードとfunction化
