+++ lib/Dialog.js
@@ -5,7 +5,7 @@ Object.defineProperty(exports, "__esModule", {
 });
 exports.Dialog = void 0;
 
-var _react = _interopRequireWildcard(require("react"));
+var _react = _interopRequireDefault(require("react"));
 
 var _Dialog = _interopRequireDefault(require("material-ui/Dialog"));
 
@@ -15,33 +15,11 @@ var _colors = require("material-ui/styles/colors");
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { "default": obj }; }
 
-function _getRequireWildcardCache() { if (typeof WeakMap !== "function") return null; var cache = new WeakMap(); _getRequireWildcardCache = function _getRequireWildcardCache() { return cache; }; return cache; }
-
-function _interopRequireWildcard(obj) { if (obj && obj.__esModule) { return obj; } if (obj === null || _typeof(obj) !== "object" && typeof obj !== "function") { return { "default": obj }; } var cache = _getRequireWildcardCache(); if (cache && cache.has(obj)) { return cache.get(obj); } var newObj = {}; var hasPropertyDescriptor = Object.defineProperty && Object.getOwnPropertyDescriptor; for (var key in obj) { if (Object.prototype.hasOwnProperty.call(obj, key)) { var desc = hasPropertyDescriptor ? Object.getOwnPropertyDescriptor(obj, key) : null; if (desc && (desc.get || desc.set)) { Object.defineProperty(newObj, key, desc); } else { newObj[key] = obj[key]; } } } newObj["default"] = obj; if (cache) { cache.set(obj, newObj); } return newObj; }
-
-function _typeof(obj) { if (typeof Symbol === "function" && typeof Symbol.iterator === "symbol") { _typeof = function _typeof(obj) { return typeof obj; }; } else { _typeof = function _typeof(obj) { return obj && typeof Symbol === "function" && obj.constructor === Symbol && obj !== Symbol.prototype ? "symbol" : typeof obj; }; } return _typeof(obj); }
-
 function _extends() { _extends = Object.assign || function (target) { for (var i = 1; i < arguments.length; i++) { var source = arguments[i]; for (var key in source) { if (Object.prototype.hasOwnProperty.call(source, key)) { target[key] = source[key]; } } } return target; }; return _extends.apply(this, arguments); }
 
 function _objectWithoutProperties(source, excluded) { if (source == null) return {}; var target = _objectWithoutPropertiesLoose(source, excluded); var key, i; if (Object.getOwnPropertySymbols) { var sourceSymbolKeys = Object.getOwnPropertySymbols(source); for (i = 0; i < sourceSymbolKeys.length; i++) { key = sourceSymbolKeys[i]; if (excluded.indexOf(key) >= 0) continue; if (!Object.prototype.propertyIsEnumerable.call(source, key)) continue; target[key] = source[key]; } } return target; }
 
 function _objectWithoutPropertiesLoose(source, excluded) { if (source == null) return {}; var target = {}; var sourceKeys = Object.keys(source); var key, i; for (i = 0; i < sourceKeys.length; i++) { key = sourceKeys[i]; if (excluded.indexOf(key) >= 0) continue; target[key] = source[key]; } return target; }
-
-function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
-
-function _defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if ("value" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } }
-
-function _createClass(Constructor, protoProps, staticProps) { if (protoProps) _defineProperties(Constructor.prototype, protoProps); if (staticProps) _defineProperties(Constructor, staticProps); return Constructor; }
-
-function _possibleConstructorReturn(self, call) { if (call && (_typeof(call) === "object" || typeof call === "function")) { return call; } return _assertThisInitialized(self); }
-
-function _assertThisInitialized(self) { if (self === void 0) { throw new ReferenceError("this hasn't been initialised - super() hasn't been called"); } return self; }
-
-function _getPrototypeOf(o) { _getPrototypeOf = Object.setPrototypeOf ? Object.getPrototypeOf : function _getPrototypeOf(o) { return o.__proto__ || Object.getPrototypeOf(o); }; return _getPrototypeOf(o); }
-
-function _inherits(subClass, superClass) { if (typeof superClass !== "function" && superClass !== null) { throw new TypeError("Super expression must either be null or a function"); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, writable: true, configurable: true } }); if (superClass) _setPrototypeOf(subClass, superClass); }
-
-function _setPrototypeOf(o, p) { _setPrototypeOf = Object.setPrototypeOf || function _setPrototypeOf(o, p) { o.__proto__ = p; return o; }; return _setPrototypeOf(o, p); }
 
 var styles = {
   root: {},
@@ -62,40 +40,22 @@ var styles = {
   }
 };
 
-var Dialog =
-/*#__PURE__*/
-function (_Component) {
-  _inherits(Dialog, _Component);
+var Dialog = function Dialog(_ref) {
+  var style = _ref.style,
+      titleStyle = _ref.titleStyle,
+      bodyStyle = _ref.bodyStyle,
+      contentStyle = _ref.contentStyle,
+      actionsContainerStyle = _ref.actionsContainerStyle,
+      children = _ref.children,
+      props = _objectWithoutProperties(_ref, ["style", "titleStyle", "bodyStyle", "contentStyle", "actionsContainerStyle", "children"]);
 
-  function Dialog() {
-    _classCallCheck(this, Dialog);
-
-    return _possibleConstructorReturn(this, _getPrototypeOf(Dialog).apply(this, arguments));
-  }
-
-  _createClass(Dialog, [{
-    key: "render",
-    value: function render() {
-      var _this$props = this.props,
-          style = _this$props.style,
-          titleStyle = _this$props.titleStyle,
-          bodyStyle = _this$props.bodyStyle,
-          contentStyle = _this$props.contentStyle,
-          actionsContainerStyle = _this$props.actionsContainerStyle,
-          children = _this$props.children,
-          props = _objectWithoutProperties(_this$props, ["style", "titleStyle", "bodyStyle", "contentStyle", "actionsContainerStyle", "children"]);
-
-      return _react["default"].createElement(_Dialog["default"], _extends({}, props, {
-        style: _lodash["default"].defaults(style, styles.root),
-        titleStyle: _lodash["default"].defaults(titleStyle, styles.title),
-        bodyStyle: _lodash["default"].defaults(bodyStyle, styles.body),
-        contentStyle: _lodash["default"].defaults(contentStyle, styles.content),
-        actionsContainerStyle: _lodash["default"].defaults(actionsContainerStyle, styles.actionsContainer)
-      }), children);
-    }
-  }]);
-
-  return Dialog;
-}(_react.Component);
+  return _react["default"].createElement(_Dialog["default"], _extends({}, props, {
+    style: _lodash["default"].defaults(style, styles.root),
+    titleStyle: _lodash["default"].defaults(titleStyle, styles.title),
+    bodyStyle: _lodash["default"].defaults(bodyStyle, styles.body),
+    contentStyle: _lodash["default"].defaults(contentStyle, styles.content),
+    actionsContainerStyle: _lodash["default"].defaults(actionsContainerStyle, styles.actionsContainer)
+  }), children);
+};
 
 exports.Dialog = Dialog;
