+++ lib/index.js
@@ -70,7 +70,7 @@ Object.defineProperty(exports, "TextField", {
   }
 });
 
-var _Theme = require("./Theme.js");
+var _Theme = require("./Theme");
 
 var _Dialog = require("./Dialog");
