+++ package.json
@@ -7,7 +7,7 @@
     "lib"
   ],
   "scripts": {
-    "build": "babel ./src -d ./lib",
+    "build": "npm run build-css && babel ./src -d ./lib",
     "build-css": "node-sass src/index.scss lib/index.css; cp src/index.scss lib/.",
     "storybook": "start-storybook"
   },
@@ -25,8 +25,8 @@
     "classnames": "^2.2.5",
     "lodash": "^4.17.2",
     "material-ui": "^0.20.2",
-    "react": "^16.3.2",
-    "react-dom": "16.3.3"
+    "react": "^15.6.2",
+    "react-dom": "^15.6.2"
   },
   "devDependencies": {
     "@babel/cli": "^7.2.3",
