+++ src/FlatButton.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import MuiFlatButton from 'material-ui/FlatButton';
+
+export const FlatButton = props => {
+  return (
+    <MuiFlatButton
+      { ...props }
+      />
+  )
+}
