+++ src/RaisedButton.js
@@ -1,0 +1,11 @@
+import React from 'react';
+import MuiRaisedButton from 'material-ui/RaisedButton';
+import _ from 'lodash';
+
+export const RaisedButton = (props) => {
+  return (
+    <MuiRaisedButton
+      { ...props }
+      />
+  )
+}
