+++ stories/Provider.stories.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider';
-import { defaultTheme } from '../src/style/Theme';
+import { defaultTheme } from '../src';
 
 const MuiDecorator = storyFn => <MuiThemeProvider muiTheme={ defaultTheme }>{storyFn()}</MuiThemeProvider>;
